enhancement: deselect users + form comps
